ref(onboarding): Split Maui onboarding docs (#102941)
Contributes to
https://linear.app/getsentry/issue/TET-864/introduce-folders-for-onboarding-platforms
